Refactored graphs to better seperate Vector and Categorical graphs
