+++ sci_analysis/graphs/__init__.py
@@ -3,4 +3,4 @@ Modules:
     graphs - graphing classes
 """
 
-# from graph import *
+from .vector import GraphHisto, GraphScatter, GraphBoxplot
