+++ sci_analysis/graphs/vector.py
@@ -1,101 +1,31 @@
-"""sci_analysis module: graph
-Classes:
-    Graph - The super class all other sci_analysis graphing classes descend from.
-    GraphHisto - Draws a histogram.
-    GraphScatter - Draws an x-by-y scatter plot.
-    GraphBoxplot - Draws box plots of the provided data as well as an optional probability plot.
-
-"""
-# from __future__ import absolute_import
-# from __future__ import print_function
-
 # matplotlib imports
 from matplotlib.pyplot import show, subplot, yticks, xlabel, ylabel, figure, setp, savefig, close, xticks, \
     subplots_adjust
 from matplotlib.gridspec import GridSpec
 
 # Numpy imports
-from numpy import polyfit, polyval, sort, arange, array, linspace, mgrid, vstack, reshape, argsort
+from numpy import polyfit, polyval, sort, arange, array, linspace, mgrid, vstack, reshape
 
 # Scipy imports
 from scipy.stats import probplot, gaussian_kde
 
 # local imports
+from .base import Graph
 from ..data import Vector, is_dict, is_group
-# TODO: Add preferences back in a future version
-# from ..preferences.preferences import GraphPreferences
-# from six.moves import range
+from ..analysis.exc import MinimumSizeError, NoDataError
 
 
-class MinimumSizeError(Exception):
-    """Thrown when the length of the Data object is less than the Graph object's _min_size property."""
-    pass
-
-
-class NoDataError(Exception):
-    """Thrown when the Data object passed to a Graph object is empty or has no graph-able data."""
-    pass
-
-
-class Graph(object):
-    """The super class all other sci_analysis graphing classes descend from.
-    Classes that descend from Graph should implement the draw method at bare minimum.
-
-    Graph members are _nrows, _ncols, _xsize, _ysize, _data, _xname and _yname. The _nrows
-    member is the number of graphs that will span vertically. The _ncols member is
-    the number of graphs that will span horizontally. The _xsize member is the horizontal
-    size of the graph area. The _ysize member is the vertical size of the graph area.
-    The _data member the data to be plotted. The _xname member is the x-axis label.
-    The _yname member is the y-axis label.
-
-    Parameters
-    ----------
-    _nrows : int, static
-        The number of graphs that will span vertically.
-    _ncols : int, static
-        The number of graphs that will span horizontally.
-    _xsize : int, static
-        The horizontal size of the graph area.
-    _ysize : int, static
-        The vertical size of the graph area.
-    _min_size : int, static
-        The minimum required length of the data to be graphed.
-    _xname : str
-        The x-axis label.
-    _yname : str
-        The y-axis label.
-    _data : Data or list(d1, d2, ..., dn)
-        The data to graph.
-
-    Returns
-    -------
-    pass
-    """
-
-    _nrows = 1
-    _ncols = 1
-    _xsize = 5
-    _ysize = 5
-    _min_size = 1
+class VectorGraph(Graph):
 
     def __init__(self, *args, **kwargs):
         """Converts the data argument to a Vector object and sets it to the Graph
         object's vector member. Sets the xname and yname arguments as the axis
         labels. The default values are "x" and "y".
-
-        Parameters
-        ----------
-        args : tuple
-            The data to be graphed.
-        kwargs : dict
-            The input parameters that control graphing behavior.
         """
 
-        self._xname = kwargs['xname'] if 'xname' in kwargs else 'x'
-        self._yname = kwargs['yname'] if 'yname' in kwargs else 'y'
-
-        if 'intersect' in kwargs:
-            # x, y = assign(args[0], args[1])
+        if 'intersect' in kwargs.keys():
+            if len(tuple(args)) != 2:
+                raise ValueError('x and y vectors must be suplied if intersect is set.')
             x, y = Vector(args[0]), Vector(args[1])
             if x is None or y is None:
                 raise NoDataError("Cannot graph because there is no data")
@@ -105,11 +35,10 @@ class Graph(object):
                 raise NoDataError("Cannot perform test because there is no data")
             if len(x) <= self._min_size or len(y) <= self._min_size:
                 raise MinimumSizeError("length of data is less than the minimum size {}".format(self._min_size))
-            self._data = [x, y]
+            super(VectorGraph, self).__init__((x, y), **kwargs)
         else:
             data = list()
             for d in args:
-                # clean = assign(d).data_prep()
                 clean = Vector(d).data_prep()
                 if clean is None:
                     data.append(None)
@@ -121,46 +50,8 @@ class Graph(object):
                 raise NoDataError("Cannot perform test because there is no data")
             if len(data) == 1:
                 data = data[0]
-            self._data = data
+            super(VectorGraph, self).__init__(data, **kwargs)
         self.draw()
-
-    @staticmethod
-    def get_color(num):
-        """Return a color based on the given num argument.
-
-        Parameters
-        ----------
-        num : int
-            A numeric value greater than zero that returns a corresponding color.
-
-        Returns
-        -------
-        color : tuple
-            A color tuple calculated from the num argument.
-        """
-        colors = [(0.0, 0.3, 0.7, 1.0),     # blue
-                  (1.0, 0.1, 0.1, 1.0),     # red
-                  (0.0, 0.7, 0.3, 1.0),     # green
-                  (1.0, 0.5, 0.0, 1.0),     # orange
-                  (0.1, 1.0, 1.0, 1.0),     # cyan
-                  (1.0, 1.0, 0.0, 1.0),     # yellow
-                  (1.0, 0.0, 1.0, 1.0),     # magenta
-                  (0.5, 0.0, 1.0, 1.0),     # purple
-                  (0.5, 1.0, 0.0, 1.0),     # light green
-                  (0.0, 0.0, 0.0, 1.0)      # black
-                  ]
-        desired_color = []
-        if num < 0:
-            num *= -1
-        floor = int(num) // len(colors)
-        remainder = int(num) % len(colors)
-        selected = colors[remainder]
-        if floor > 0:
-            for value in selected:
-                desired_color.append(value / (2.0 * floor) + 0.4)
-            return tuple(desired_color)
-        else:
-            return selected
 
     def draw(self):
         """
@@ -169,7 +60,7 @@ class Graph(object):
         raise NotImplementedError
 
 
-class GraphHisto(Graph):
+class GraphHisto(VectorGraph):
     """Draws a histogram.
 
     New class members are bins, color and box_plot. The bins member is the number
@@ -366,7 +257,7 @@ class GraphHisto(Graph):
         pass
 
 
-class GraphScatter(Graph):
+class GraphScatter(VectorGraph):
     """Draws an x-by-y scatter plot.
 
     Unique class members are fit and style. The fit member is a boolean flag for
@@ -537,7 +428,7 @@ class GraphScatter(Graph):
         pass
 
 
-class GraphBoxplot(Graph):
+class GraphBoxplot(VectorGraph):
     """Draws box plots of the provided data as well as an optional probability plot.
 
     Unique class members are groups, nqp and prob. The groups member is a list of
